enhancement(Overview): Add slimmed down version of Host Overview as OrgOverview
