+++ lang/es.json
@@ -3374,9 +3374,6 @@
   "settings.general": "General",
   "Settings.Updated": "Configuración actualizada.",
   "Settings.Updated.Fail": "La actualización ha fallado.",
-  "SetupGuide.HideSetupGuide": "Hide setup guide",
-  "SetupGuide.ShowSetupGuide": "Show setup guide",
-  "SetupGuide.Title": "Setup guide",
   "Sg4yBm": "<Individual></Individual> rechazó la solicitud de tarjeta virtual a <Account></Account>",
   "sgLhGq": "Más información sobre cómo funciona la plataforma",
   "sgVok2": "El tipo de cuenta de la transacción contraria (persona, organización, Colectivo, Anfitrión, fondo, etc.)",
