+++ lang/fr.json
@@ -3374,9 +3374,6 @@
   "settings.general": "Général",
   "Settings.Updated": "Paramètres mis à jour.",
   "Settings.Updated.Fail": "Échec de la mise à jour.",
-  "SetupGuide.HideSetupGuide": "Hide setup guide",
-  "SetupGuide.ShowSetupGuide": "Show setup guide",
-  "SetupGuide.Title": "Setup guide",
   "Sg4yBm": "<Individual></Individual> a rejeté la demande de carte virtuelle à <Account></Account>",
   "sgLhGq": "En savoir plus sur le fonctionnement de la plateforme",
   "sgVok2": "Le type de compte de contrepartie (Personne, Organisation, Collectif, Hôte, Fonds, etc.)",
