+++ lang/ja.json
@@ -3374,9 +3374,6 @@
   "settings.general": "General",
   "Settings.Updated": "Settings updated.",
   "Settings.Updated.Fail": "Update failed.",
-  "SetupGuide.HideSetupGuide": "Hide setup guide",
-  "SetupGuide.ShowSetupGuide": "Show setup guide",
-  "SetupGuide.Title": "Setup guide",
   "Sg4yBm": "<Individual></Individual> rejected request for virtual card to <Account></Account>",
   "sgLhGq": "プラットフォームの機能の詳細についてはこちら",
   "sgVok2": "The type of the transaction opposite account (person, organization, collective, host, fund, etc.)",
