+++ lang/pl.json
@@ -3374,9 +3374,6 @@
   "settings.general": "General",
   "Settings.Updated": "Zaktualizowano ustawienia.",
   "Settings.Updated.Fail": "Aktualizacja nie powiodła się.",
-  "SetupGuide.HideSetupGuide": "Hide setup guide",
-  "SetupGuide.ShowSetupGuide": "Show setup guide",
-  "SetupGuide.Title": "Setup guide",
   "Sg4yBm": "<Individual></Individual> rejected request for virtual card to <Account></Account>",
   "sgLhGq": "Dowiedz się więcej o działaniu platformy",
   "sgVok2": "The type of the transaction opposite account (person, organization, collective, host, fund, etc.)",
