+++ lang/sv-SE.json
@@ -3374,9 +3374,6 @@
   "settings.general": "General",
   "Settings.Updated": "Inställningar är uppdaterade.",
   "Settings.Updated.Fail": "Uppdatering misslyckades.",
-  "SetupGuide.HideSetupGuide": "Hide setup guide",
-  "SetupGuide.ShowSetupGuide": "Show setup guide",
-  "SetupGuide.Title": "Setup guide",
   "Sg4yBm": "<Individual></Individual> rejected request for virtual card to <Account></Account>",
   "sgLhGq": "Läs mer om hur plattformen fungerar",
   "sgVok2": "The type of the transaction opposite account (person, organization, collective, host, fund, etc.)",
