+++ lang/uk.json
@@ -3374,9 +3374,6 @@
   "settings.general": "General",
   "Settings.Updated": "Налаштування оновлено.",
   "Settings.Updated.Fail": "Не вдалося оновити.",
-  "SetupGuide.HideSetupGuide": "Hide setup guide",
-  "SetupGuide.ShowSetupGuide": "Show setup guide",
-  "SetupGuide.Title": "Setup guide",
   "Sg4yBm": "<Individual></Individual> rejected request for virtual card to <Account></Account>",
   "sgLhGq": "Дізнатися більше про роботу платформи",
   "sgVok2": "The type of the transaction opposite account (person, organization, collective, host, fund, etc.)",
