+++ lang/zh.json
@@ -3374,9 +3374,6 @@
   "settings.general": "General",
   "Settings.Updated": "设置已更新。",
   "Settings.Updated.Fail": "更新失败。",
-  "SetupGuide.HideSetupGuide": "Hide setup guide",
-  "SetupGuide.ShowSetupGuide": "Show setup guide",
-  "SetupGuide.Title": "Setup guide",
   "Sg4yBm": "<Individual></Individual> rejected request for virtual card to <Account></Account>",
   "sgLhGq": "了解更多关于平台如何工作",
   "sgVok2": "The type of the transaction opposite account (person, organization, collective, host, fund, etc.)",
